TS Move common types to core package #9779
Moved DSL Queries to core.
